Handheld Showcase
Provided the ability to press the Showcase Key and showcase the currently held item, if it is cobbleTCG related!
